# server/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Foodsharing EinAb API
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/teacher/me:
    get:
      tags: [teacher]
      operationId: getTeacherMe
      responses:
        '200':
          description: Teacher profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherSelfResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [teacher]
      operationId: updateTeacherMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeacherMeRequest'
      responses:
        '200':
          description: Updated teacher profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherSelfResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teacher/bezirke/{bezirkSlug}/einabs:
    get:
      tags: [teacher]
      operationId: getTeacherEinAbs
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Teacher EinAbs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherEinAbListResponse'
    post:
      tags: [teacher]
      operationId: createTeacherEinAb
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertEinAbRequest'
      responses:
        '201':
          description: EinAb created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherEinAbResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teacher/bezirke/{bezirkSlug}/einabs/{einAbId}:
    put:
      tags: [teacher]
      operationId: updateTeacherEinAb
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - $ref: '#/components/parameters/EinAbId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertEinAbRequest'
      responses:
        '200':
          description: EinAb updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherEinAbResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [teacher]
      operationId: deleteTeacherEinAb
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - $ref: '#/components/parameters/EinAbId'
      responses:
        '204':
          description: EinAb deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/teacher/bezirke/{bezirkSlug}/bookings:
    get:
      tags: [teacher]
      operationId: getTeacherBookings
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Bookings for teacher slots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
  /api/teacher/bezirke/{bezirkSlug}/bookings/{bookingUserId}/comments:
    get:
      tags: [teacher]
      operationId: getTeacherBookingComments
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - name: bookingUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingCommentListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [teacher]
      operationId: addTeacherBookingComment
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - name: bookingUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingCommentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/teacher/ical-candidates:
    get:
      tags: [teacher]
      operationId: getTeacherIcalCandidates
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Candidate iCal events for the teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IcalCandidateListResponse'
  /api/teacher/bezirke/{bezirkSlug}/slots/{slotId}/booking:
    delete:
      tags: [teacher]
      operationId: cancelTeacherSlotBooking
      parameters:
        - $ref: '#/components/parameters/BezirkSlug'
        - $ref: '#/components/parameters/SlotId'
      responses:
        '200':
          description: Slot booking cancelled by teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotResponse'
...
components:
  parameters:
    BezirkSlug:
      name: bezirkSlug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9-]+$'
    SlotId:
      name: slotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EinAbId:
      name: einAbId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [message, status, code]
      properties:
        status:
          type: integer
        message:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'
        details:
          type: array
          items:
            type: string
    ErrorCode:
      type: string
      enum: [AUTHENTICATION_REQUIRED, ACCESS_DENIED, VALIDATION_FAILED, RESOURCE_NOT_FOUND, UNEXPECTED_ERROR, BEZIRK_NOT_FOUND, USER_BEZIRK_MISMATCH, NOTIFICATION_SUBSCRIPTION_BEZIRK_MISMATCH, INVALID_LOGIN_TOKEN, LOGIN_TOKEN_EXPIRED_OR_USED, ACCOUNT_NOT_FOUND, INVALID_UNSUBSCRIBE_TOKEN, TEACHER_ACCOUNT_REQUIRED, TEACHER_FOODSHARING_ID_ALREADY_EXISTS, TEACHER_NOT_FOUND, TEACHER_INACTIVE, BOOKING_USER_DISABLED, BOOKING_USER_NOT_FOUND, SLOT_NOT_FOUND, BOOKING_NOT_FOUND, SLOT_NOT_AVAILABLE, USER_ALREADY_BOOKED_WITH_TEACHER, USER_ALREADY_BOOKED_IN_CATEGORY, USER_BOOKING_LIMIT_REACHED, MINIMUM_PICKUP_COUNT_NOT_REACHED, ONLY_OWN_BOOKINGS_CANCELLABLE, ONLY_BOOKED_APPOINTMENTS_CANCELLABLE, EINAB_NOT_FOUND, ONLY_OWN_EINABS_MANAGEABLE, SLOT_COUNT_OUT_OF_RANGE, SLOT_COUNT_REDUCTION_NOT_SUPPORTED, BOOKED_SLOTS_PREVENT_DELETE, PUBLIC_LOCATION_REQUIRED, DEFAULT_ADMIN_CANNOT_BE_CHANGED, INVALID_BOOKING_CONFIRMATION_TOKEN, BOOKING_CONFIRMATION_EXPIRED]
    BezirkResponse:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
...
    Language:
      type: string
      enum: [de, en, gws]
    EinAbCategory:
      type: string
      enum: [SUPERMARKET, TAKEOUT, MARKET, BAKERY, RESTAURANT, FAIRTEILER_CLEANING]
    SlotStatus:
      type: string
      enum: [AVAILABLE, PENDING_CONFIRMATION, BOOKED, DONE, CANCELLED]
...
    TeacherResponse:
      type: object
      required: [id, foodsharingId, name, active, language, createdAt, updatedAt, canGiveEinAbs, canManageUsers, canUseAutomations, canSeeUserPickupCountGrouping, canUseAutomationSlotApproval, canUseAutomationRequestApproval, canUseAutomationOpenSlotAdvertising, canSeeAllAutomationDecisions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        foodsharingId:
          type: string
        name:
          type: string
        phoneNumber:
          type: string
        icalLink:
          type: string
        active:
          type: boolean
        canGiveEinAbs:
          type: boolean
        canManageUsers:
          type: boolean
        canUseAutomations:
          type: boolean
        canSeeUserPickupCountGrouping:
          type: boolean
        canUseAutomationSlotApproval:
          type: boolean
        canUseAutomationRequestApproval:
          type: boolean
        canUseAutomationOpenSlotAdvertising:
          type: boolean
        canSeeAllAutomationDecisions:
          type: boolean
        wantsToBeTeacher:
          type: boolean
        language:
          $ref: '#/components/schemas/Language'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
...
    TeacherSelfResponse:
      allOf:
        - $ref: '#/components/schemas/TeacherResponse'
        - type: object
          required: [icalCandidates]
          properties:
            icalCandidates:
              type: array
              items:
                $ref: '#/components/schemas/IcalCandidate'
    UpdateTeacherMeRequest:
      type: object
      required: [phoneNumber, language]
      properties:
        phoneNumber:
          type: string
          minLength: 3
          maxLength: 50
        icalLink:
          type: string
          maxLength: 2000
        language:
          $ref: '#/components/schemas/Language'
    BookingUserResponse:
      type: object
      required: [id, name, foodsharingId, language, active, createdAt, updatedAt, canGiveEinAbs, canManageUsers, canUseAutomations, canSeeUserPickupCountGrouping, canUseAutomationSlotApproval, canUseAutomationRequestApproval, canUseAutomationOpenSlotAdvertising, canSeeAllAutomationDecisions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        foodsharingId:
          type: string
        phoneNumber:
          type: string
        language:
          $ref: '#/components/schemas/Language'
        active:
          type: boolean
        canGiveEinAbs:
          type: boolean
        canManageUsers:
          type: boolean
        canUseAutomations:
          type: boolean
        canSeeUserPickupCountGrouping:
          type: boolean
        canUseAutomationSlotApproval:
          type: boolean
        canUseAutomationRequestApproval:
          type: boolean
        canUseAutomationOpenSlotAdvertising:
          type: boolean
        canSeeAllAutomationDecisions:
          type: boolean
        wantsToBeTeacher:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
...
    EinAbResponse:
      type: object
      required: [id, bezirk, category, startDateTime, publicLocation, teacher, visitFairteiler, slotCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
        category:
          $ref: '#/components/schemas/EinAbCategory'
        startDateTime:
          type: string
          format: date-time
        location:
          type: string
        publicLocation:
          type: string
        whatToBring:
          type: string
        hint:
          type: string
        teacher:
          $ref: '#/components/schemas/TeacherResponse'
        visitFairteiler:
          type: boolean
        slotCount:
          type: integer
        minimumPickupCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SlotResponse:
      type: object
      required: [id, einAbId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        einAbId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SlotStatus'
        bookedAt:
          type: string
          format: date-time
        doneAt:
          type: string
          format: date-time
        bookingUser:
          $ref: '#/components/schemas/BookingUserResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BookingDetailResponse:
      type: object
      required: [slotId, einAbId, bezirk, category, startDateTime, teacherName, teacherPhoneNumber, status, bookingUser]
      properties:
        slotId:
          type: string
          format: uuid
        einAbId:
          type: string
          format: uuid
        bezirk:
          $ref: '#/components/schemas/BezirkResponse'
        category:
          $ref: '#/components/schemas/EinAbCategory'
        startDateTime:
          type: string
          format: date-time
        location:
          type: string
        publicLocation:
          type: string
        whatToBring:
          type: string
        hint:
          type: string
        teacherName:
          type: string
        teacherPhoneNumber:
          type: string
        visitFairteiler:
          type: boolean
        status:
          $ref: '#/components/schemas/SlotStatus'
        bookingUser:
          $ref: '#/components/schemas/BookingUserResponse'
    BookingListResponse:
      type: object
      required: [bookings, page, size, totalElements, totalPages]
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/BookingDetailResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    TeacherEinAbResponse:
      allOf:
        - $ref: '#/components/schemas/EinAbResponse'
        - type: object
          required: [slots]
          properties:
            slots:
              type: array
              items:
                $ref: '#/components/schemas/SlotResponse'
    TeacherEinAbListResponse:
      type: object
      required: [einAbs, page, size, totalElements, totalPages]
      properties:
        einAbs:
          type: array
          items:
            $ref: '#/components/schemas/TeacherEinAbResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    UpsertEinAbRequest:
      type: object
      required: [category, startDateTime, publicLocation, visitFairteiler, slotCount]
      properties:
        category:
          $ref: '#/components/schemas/EinAbCategory'
        startDateTime:
          type: string
          format: date-time
        location:
          type: string
          maxLength: 500
        publicLocation:
          type: string
          minLength: 1
          maxLength: 500
        whatToBring:
          type: string
          maxLength: 4000
        hint:
          type: string
          maxLength: 4000
        visitFairteiler:
          type: boolean
        slotCount:
          type: integer
          minimum: 1
          maximum: 3
        minimumPickupCount:
          type: integer
          minimum: 0
    CreateBookingCommentRequest:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
          minLength: 1
          maxLength: 4000
    BookingCommentResponse:
      type: object
      required: [id, bookingUserId, teacherId, teacherName, comment, createdAt]
      properties:
        id:
          type: string
          format: uuid
        bookingUserId:
          type: string
          format: uuid
        teacherId:
          type: string
          format: uuid
        teacherName:
          type: string
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    BookingCommentListResponse:
      type: object
      required: [comments]
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/BookingCommentResponse'
    IcalCandidate:
      type: object
      required: [startDateTime, summary]
      properties:
        startDateTime:
          type: string
          format: date-time
        summary:
          type: string
        location:
          type: string
    IcalCandidateListResponse:
      type: object
      required: [candidates, page, size, totalElements, totalPages]
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/IcalCandidate'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean